restructure main into server and client binaries
