+++ cmd/meshd-client/client.go
@@ -1,0 +1,91 @@
+package main
+
+import (
+	"crypto/tls"
+	"flag"
+	"log"
+	"net/url"
+	"os"
+	"os/signal"
+	"time"
+
+	"github.com/gorilla/websocket"
+)
+
+var clientTlsAddr = flag.String("clientTlsAddr", "100.119.2.72:4443", "https client address")
+
+func main() {
+	flag.Parse()
+	log.SetFlags(0)
+
+	if err := run(); err != nil {
+		log.Fatalln(err)
+	}
+}
+
+func run() error {
+	interrupt := make(chan os.Signal, 1)
+	signal.Notify(interrupt, os.Interrupt)
+
+	u := url.URL{Scheme: "wss", Host: *clientTlsAddr, Path: "/socket"}
+	log.Printf("connecting to %s", u.String())
+	dialer := websocket.DefaultDialer
+	tlsConfig := tls.Config{
+		NextProtos:               nil,
+		ServerName:               "mesh",
+		InsecureSkipVerify:       true,
+		CipherSuites:             []uint16{tls.TLS_AES_256_GCM_SHA384},
+		PreferServerCipherSuites: true,
+		MinVersion:               tls.VersionTLS13,
+		MaxVersion:               tls.VersionTLS13,
+	}
+	dialer.TLSClientConfig = &tlsConfig
+	c, _, err := dialer.Dial(u.String(), nil)
+	if err != nil {
+		log.Fatal("dial:", err)
+	}
+	defer c.Close()
+
+	done := make(chan struct{})
+
+	go func() {
+		defer close(done)
+		for {
+			_, message, err := c.ReadMessage()
+			if err != nil {
+				log.Println("read:", err)
+				return
+			}
+			log.Printf("recv: %s", message)
+		}
+	}()
+
+	ticker := time.NewTicker(time.Second)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-done:
+			return nil
+		case t := <-ticker.C:
+			err := c.WriteMessage(websocket.TextMessage, []byte(t.String()))
+			if err != nil {
+				return err
+			}
+		case <-interrupt:
+			log.Println("interrupt")
+
+			// Cleanly close the connection by sending a close message and then
+			// waiting (with timeout) for the server to close the connection.
+			err := c.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseNormalClosure, ""))
+			if err != nil {
+				return err
+			}
+			select {
+			case <-done:
+			case <-time.After(time.Second):
+			}
+			return nil
+		}
+	}
+}
