+++ meshd_test.go
@@ -1,4 +1,4 @@
-package main
+package meshd_test
 
 import "testing"
 
@@ -6,11 +6,10 @@ func Test_main(t *testing.T) {
 	tests := []struct {
 		name string
 	}{
-		// TODO: Add test cases
+		// TODO: Use ginkgo for BDD testing
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			main()
 		})
 	}
 }
