+++ websocket.go
@@ -1,15 +1,15 @@
-package main
+package meshd
 
 // Works for websocket client and server
 type Transport interface {
-	Open()        // Client will connect, server will listen
-	Close()       // Client will disconnect, server will stop listening and disconnect all clients
-	OnConnect()   // When we connect or re-connect
-    OnDisconnect() // Handle disconnection
-	OnMessage()   // When we receive a new message from the other endpoint
-	OnError()     // When we have an error
+	Open()         // Client will connect, server will listen
+	Close()        // Client will disconnect, server will stop listening and disconnect all clients
+	OnConnect()    // When we connect or re-connect
+	OnDisconnect() // Handle disconnection
+	OnMessage()    // When we receive a new message from the other endpoint
+	OnError()      // When we have an error
 
-    SendMessage() // Send a new message to the other endpoint
+	SendMessage() // Send a new message to the other endpoint
 	SendReply()   // Send a response to a message we sent
 }
 
@@ -78,8 +78,6 @@ func (wc WebsocketClient) OnError() {
 }
 
 type EmailAgent interface {
-    // TODO: Some interface for polling via IMAP or POP3 or ???
-    SendPlainTextEmail(targetEmail string, subject string, body string)
-    
+	// TODO: Some interface for polling via IMAP or POP3 or ???
+	SendPlainTextEmail(targetEmail string, subject string, body string)
 }
-
